pyproject.toml
[project]
name = "dtc-zc-data-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "cassandra-driver>=3.29.3",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
]
